moved Estimates class and related methods on a different file
